IndustrialNoiseAI/requirements.txt
flask
numpy
pillow
scipy
librosa
matplotlib
